refactor: mudança e exception de pacotes para seguir a arquitetura haxagonal
