+++ financas_bot_telegram/src/main/java/br/com/satyan/stering/saita/financasbottelegram/adapters/in/telegram/exception/InvalidCaptionException.java
@@ -1,4 +1,4 @@
-package br.com.satyan.stering.saita.financasbottelegram.application.exceptions;
+package br.com.satyan.stering.saita.financasbottelegram.adapters.in.telegram.exception;
 
 public class InvalidCaptionException extends RuntimeException {
